Сделал суперкласс для всех моих CursorLoader, вынес в него фильтры
